requirements.txt
python-telegram-bot==20.3
requests
colorama
beautifulsoup4
lxml
tqdm